Move JWE algorithms to JWA as well
This keeps all JWA defined algorithms in the same module and make all
algorithms use a common way of instantiating and handling crypto

Closes #57
